+++ packages/harness-tester/src/harnesses.ts
@@ -1,5 +1,5 @@
-import { CodexTuiHarness } from "./harness/codex/codexTui.js";
-import { ClaudeTuiHarness } from "./harness/claude/claudeTui.js";
+import { CodexTuiHarness } from "./harness/codex/codexHarness.js";
+import { ClaudeTuiHarness } from "./harness/claude/claudeHarness.js";
 import type { AgentHarness } from "./harness/types.js";
 
 export const AGENT_HARNESSES: readonly (new () => AgentHarness)[] = [CodexTuiHarness, ClaudeTuiHarness];
