+++ packages/harness-tester/src/index.ts
@@ -1,5 +1,5 @@
-import { ClaudeTuiHarness } from "./harness/claude/claudeTui.js";
-import { CodexTuiHarness } from "./harness/codex/codexTui.js";
+import { ClaudeTuiHarness } from "./harness/claude/claudeHarness.js";
+import { CodexTuiHarness } from "./harness/codex/codexHarness.js";
 
 export { useAgent, type AgentContext } from "./useAgent.js";
 export { AGENT_HARNESSES } from "./harnesses.js";
